refactor: separate inventory validation and harden SVG parsing
